requirements.txt
PyPDF2>=3.0.0
tabula-py>=2.9.0
pandas>=2.0.0
